requirements.txt
aiohttp==3.5.1
apscheduler==3.5.3
Flask==1.0.2
flask-allows==0.7.1
flask-apispec==0.8.0
Flask-INIConfig==0.1.0
Flask-JSGlue==0.3.1
flask-ldap3-login==0.9.16
Flask-Login==0.4.1
flask-marshmallow==0.9.0
Flask-SQLAlchemy==2.3.2
humanize==0.5.1
marshmallow==3.0.0rc5
marshmallow-sqlalchemy==0.15.0
pbr==5.1.1
psycopg2-binary==2.8.2
PTable==0.9.2
python-dateutil==2.7.5
stevedore==1.30.0
uWSGI==2.0.17.1
webargs==5.1.3
